requirements/base.txt
fastapi==0.112.0
uvicorn==0.30.5
psycopg2==2.9.9
python-dotenv==1.0.1
argon2-cffi==23.1.0
PyJWT==2.9.0
email_validator==2.2.0
